[1.4][sfSympalPlugin][1.0] Major refactoring of the frontend slot editing - still a work in progress.
This creates a new slots.css and slots.js file to handle the new editing of the slots on the frontend. The slots will now be
modified individually by clicking an edit button. This will bring up the form in a fancybox (I might allow for inline-editing later
as well).

There's still a lot more to be done - the save and updating mechanism for the slots is not there yet and a lot of old code still needs to be ripped out.