[1.4][sfSympalPlugin][1.0] Added a refactored, simpler version of the get_sympal_content_slot() helper function.
This was setup in such a way as to keep backwards compatibility.

The new function only has one required argument - the slot name. The only other argument is an array which handles everything else.
An addition to the app.yml configuration was made called content_slots, which goes beneath each individual content type config. This
allows you to define slot option defaults for all slot names of a given content type. These can be overridden via the second option
of the get_sympal_content_slot() helper.