[1.4][sfSympalPlugin][1.0] Refactoring a lot of the markup, class names and id attributes related to editing slots.
I'm trying to make things more consistent. The wrapper has a unique id based on the slot, and then everything keys
off of that element. I've removed other excess settings of the slot id in the dom. I've also changed the class names
of everything to conform to the sympal standard of prefixing with the word sympal.

For in-place, the edit and cancel now work, though the cancel doesn't yet leave you with updated content.
For the popup, only the edit works (cancel won't do anything yet).