[1.4][sfSympalPlugin][1.0] Fixing a bug when you edit in popup/fancybox land with referencing dom parents.
When you edit via fancybox, it literally picks up the content you're rendering and moves it. So, the slot form
itself, which normally lives inside .sympal_slot_wrapper is moved out of it. So, many things need to be careful
not to key off of the .sympal_slot_wrapper in certain cases.

I've added an id to the actual form element and triggered off of that. The css was also changed to not anchor on
the .sympal_slot_wrapper when referencing the form.