[1.4][sfSympalPlugin][1.0] Fixing close behavior for popup editing.
This addresses the selectors used when cleaning up the dom when fancybox closes. This is the same issue as the previous
commit where the fact that fancybox moves our form outside of .sympal_slot_wrapper means that we need to target in a different
way. Fortunately, this is easy since only one #fancybox-inner can be open at any given time, so we target it.