[1.4][sfSympalPlugin][1.0] Partial removal of the slot rendering system (to be replaced by the transformer).
Most of the slot rendering system was removed. A slot renderer is still used (sfSympalContentSlotRenderer), but is basically just used
to retrieve the correct raw value. Any transformations (e.g. markdown to html) will be done via the transformers.

The still unfinished unit test was updated. The replacer listener on the sympal.content_renderer.filter_slot_content was removed
and will be put back as a transformer.