[1.4][sfSympalPlugin][1.0] Adding getValueForRendering() method to sfSympalContentSlot.
This completely replaces the old "slot renderer". Now, when a slot is rendered, an getXXXSlotValue() method is looked for on the
content object. If it exists, it is called, else the raw value is returned. In either case, that value is then passed through the transformer chain.

The entire process looks like this.
  1) The slot looks for the getXXXSlotValue() method on the content object (where XXX is the camelized slot name)
  2) If it exists, that method's return value is used for the slot. Else, the raw value is used.
  3) The resulting value, regardless of method, is passed down the transformer chain.

This initial "transformation" is intended to make some large data change to your slot's value, whereas the transformers are primarily intended to
act upon and modify text.

This initial getXXXSlotValue() means of modifying a value COULD be done with a transformation. However, this allows you to make modifications to
a slot on a field-by-field basis. I think this is the flexibility we'll want.

Right now, this is used, for example, to render the created_by_id as a string username and the date_published as a formatted date.