[1.4][sfSympalPlugin][1.0] Fixing bug where column slots didn't respect widget and validator of slot type.
The downside is that I can't think of a great way to do this in the sfSympalContent or content type forms themselves. Since the widget/validator
switching logic doesn't live there, the field will still not respect slot type when editing on the backend. This is specifically because I can't
think of a way to add it to the content type as there are many content types of many different classes.

I still need to create a "column" slot type where the widget/validator from the form is respected.