[1.4][sfSympalPlugin][1.0] Fixed popup editing bug with markitup.
With fancybox, any javascript that is included in the content to be fancybox'ed is re-executed since the entire content
is actually moved. This meant that the markitup code was run twice, creating a markitup editor inside an editor. With this
fix, the markitup code is only applied if it hasn't been already applied before.