[1.4][sfSympalPlugin][1.0] Making all slots use a sfSympalContentSlot form.
This is a significant philosophical change. Column slots hold their data on either the sfSympalContent or the content object's model,
but have a sfSympalContentSlot record to represent them. They are basically rendered via the directions in sfSympalContentSlot, but
their value is saved at the last second to the actual content or content type record (instead of on the slot table).

Slots are very powerful, but it leaves us in a strange spot. When we are editing a column slot, should we be rendering a form that
ultimate extends sfSympalContentSlot or should we be rendering the form based on the true column (a form extending either sfSympalContentForm
or the content type's form).

For consistency, I've decided to try to push everything through sfSympalContentSlot. The fact that a slot is a "column" slot or just a normal
slot should, for the most part, be invisible to the developer. The developer just wants to render content, and the slot is an easy way of
essentially "extending" the fields on the content type object. For that reason, I think that a slot should always be rendered with a base
form class of sfSympalContentSlotForm.

This makes for a few problems, however. Namely, in a lot of cases, you'll want your column slot to render with the widget and validator that
you setup (or was setup automatically) for that field in the actual sfSympalContentForm or content type's form. In these cases, setting the
slot's type to "Column" will result in this behavior. We will accomplish this by actually instantiating a content or content type form and
getting the widget and validator out of it. This will be the default type for column slots.