[1.4][sfSympalPlugin][1.0] Adding functionality to call sub-function for a particular slot type if it exists.
In other words, if you have your own slot type, you can create a jquery function that will be called when the form
for that slot type is loaded.

As an example, this is done for markdown. The widget has had all of the inline javascript removed, though I still need
to think of a better way to bring in the external javascripts and css needed.