[1.0][sfEditorPlugin] Got more exact with the non-edit controls for frontend slot editing
There are seveal jquery events that occur when a slot is on the page in non-edit mode. These include
a button that displays on hover, a transparency change, and a double-click to edit.

But, as soon as inline edit mode is enabled, all these events need to disappear. So, we now keep track
of all of these events in an array. Whenever the non-edit-mode is enabled, we bind them. When the slot
is changed to edit-mode, we unbind them. This refs #1.

I also fixed the cancel functionality. It was not going through the closeEditor method - it now is. I
also changed the order in which things occur: everything gets set back up before the user sees the
edit form go away.