[1.0][sfSympalEditorPlugin] Cleaning up how the editor is closed, preventing multiple ajax calls to view content.
In the case of fancybox, if you click the X in the corner, it fires an event that calls the closeEditor() method.
This, in turn, tells fancybox to close again, which again calls closeEditor(). This cylce is prevented by unsetting
the editor data the first time closeEditor() is called. On the second time through, it just quits.