[1.0][sfSympalMenuPlugin] Failing gracefully if a menu item was given an invalid route.
There still may be more to do on this (so that the user is not unaware), but now there won't be an error thrown
and the fact that the url generation failed is logged. The raw url (e.g. this/that) entered by the user is returned
as the url.