[1.0][sfSympalPlugin] Removing unnecessary _doctrineManager property.
The Doctrine_Manager instance can always be returned statically (and is retrieved statically here), so this seems like an unnecessary helper.
I'm really trying to remove the magic or blurring of lines. There IS a way to get the manager out of Doctrine - I'd rather leave that alone
then start confusing what's what.