[1.0][sfSympalPlugin] Removing the base Doctrine_Table class - sfSympalDoctrineTable.
This class didn't do much, and if it's necessary, I'd rather do it via a Doctrine template. This to me is too much magic. If sympal were installed into a project
that was overriding the Doctrine_Table class already, this would replace it (a rare occurrence probably, but still).

This will require all generated table classes to change their base class from sfSympalDoctrineTable to Doctrine_Table