[1.0][sfSympalPlugin] Wrapping a few response.filter_content listeners to be safe.
It turns out that when rendering the exception in dev mode, the response.filter_content event is thrown. If a listener on that throws an exception, it will just be swallowed and nothing will be printed.

There may be a good reason for this, not sure. Ticket url: http://trac.symfony-project.org/ticket/8496