[1.0][sfSympalCMFPlugin] Renaming the sfSympalContentPlugin to sfSympalCMFPlugin.
There are a bunch of things that Sympal can do for you that are CMF-like, but aren't strictly related to the Content model. At this point, I think it's more helpful to not split hairs too much, and to create a catch-all plugin (sfSympalCMFPlugin) to house all of the CMF-related functionality. Obviously, the goal will be to make this plugin house as little as possible to keep things separated.