[1.0][sfSympalCMFPlugin] Moving the sympal context method_not_found listener out of the service.
Having listners inside a service is a bad idea since services are lazy-loaded. So, if you add a listener from inside your service and that service hasn't been asked for yet (and so hasn't been instantiated), your listener won't be registered.

A better way is to register a listener in your configuration. If the event is thrown, that listener will ask for the service (and it will be created right then if necessary).