[1.0][sfSympalPlugin] Major change in slot form philosophy - making everything extend its true base model.
This is a reversal of most of the idea behind revision 63cfd42d. That revision made all slots ultimately extend sfSympalSlotForm regardless of whether or not they were a column slot or not.

The problem was with I18N. While we can control where a normal form saves by overriding the setValue() on sfSympalContentSlot (forcing it to write to the content record instead), we can't do that for sfSympalContentSlotTranslation, which doesn't have a generated model file.

So, the reversal is to give each slot form its correct parent form. If it's a column slot, it may now have a parent sfSympalPageForm, for example.

The problem before was how to assign the correct widget & validator to a column slot field based on that slot's type. This was solved by running the field through sfSympalFormToolkit::changeContentSlotValueWidget() inside sfSympalContentSlot::_getContentSlotColumnForm(). This appears to have the desired effect.