[1.0][sfSympalPlugin] Changing the identification of content type models in the config from their slug to their name.
This will mean that multiple content type records will share the same base config. But, I think this is good. You need a programmer to create a new
content type model. But you only need the CMS to create multiple content type records based off of the same model. Since the configuration must be
entered by a programmer, it only makes sense that creating a new content type record in the CMS based off of an existing model would use the existing
config for that model.

This will break custom content type configurations - an upgrade script will be needed.