[1.0][sfSympalCMFPlugin] Rework of the process by which we determine the template to render for a content record.
This removes the default_content_template key both globally, and on individual content types. This is replaced by relying more on the default_view of each content type. Each content has a content type, and each content type must have a default_view template name. So, for each content type, we have an absolute backup with needing another absolute default template.

In doing this, the sfSympalContent::getTemplateToRenderWith() was reworked to behave more cleanly. Before, if a template name wasn't recognized in the template config, the template name itself would try to render (e.g. my_template), which of course is not enough information to actually render. This threw a very unhelpful error.