[1.0][sfSympalEditorPlugin] Correcting how the "popup" editor displays.
Instead of having the popup try to figure out its size, we'll just always popup with a set size that is somewhat large. Then the textarea and input elements are designed to fill up that area.

This closes #38 where popup slot editors would be very small, because the very element that would force the fancybox to grow larger, was set to have its width as a percentage of its fancy box container. The result was a small fancybox.