Updating config for new sfThemePlugin config and removing sfJqueryReloadedPlugin and most references to it.
The first part is obvious - just some moving of values in app.yml for their new locations.

The removal of the jQuery reloaded plugin was in an attempt to take more control of how css and js is loaded.

I've begun exercising the already-built assets_path config that allows you to specify a css file, and then map it to a file in the config.

But, instead of pointing to real file paths and allowing the developer to re-map those assets later, I'm pointing simply to the common file names.
I think this is going to be very clean overall and is mirrored off of diem and also (though not exactly the same) sfAssetsManagerPlugin.