fix bug where admin/sites action lists the current site everywhere
I've spent a whole day hunting for that bug.

There is a listener in Sympal, which listens template.filter_parameters event.
Besides the other it writes $parameters['sf_sympal_site'] variable to make it available from view.
This variable can be used to retrieve data about current site from view.
But, it does so without check of existance such variable, which leads to the following issue.

Problem:
admin/sites shows current site for every position in list

How it looks in the code:
When symfony gets list of sites to display it returns a pager and registers it for view as $pager variable.
Then the view iterates over the pager and tries to display content.

It uses
include_partial('sympal_sites/list_td_batch_actions', array('sf_sympal_site' => $sf_sympal_site, 'helper' => $helper))

which means that symfony will execute sfView for partial with 'sf_sympal_site' parameter.
That's where is a clue for the bug.

sfView triggers event template.filter_parameters which allows plugins modify variables which will be passed to view.
One of such listeners is sfSympalCMFPluginConfiguration::listenTemplateFilterParameters()

It designed to add current site to view context, so as view user could retrieve info about the current site.
But it does so always, overwriting any existing values.

which leads to situation where are only current site enties in the resulted list