[Editor] Making the Slot Renderer more flexible in how it renders slots.
The renderer either renders a raw slot, or a slot "editor", which is the slot wrapped in markup needed to have an inline editor attached to it.

Before, this switch was entirely determined by whether or not the user was logged in, ajax request, if there is a bound Content object etc. This is still the default behavior, but you can now force the renderer one way or another.

This is useful in the slot saving modules where we may want to refresh the slot editor completely or just return the raw value.

Also removed the requirement for there to be a bound Content object to load the editor. This method is only used by the slot renderer, and if we're rendering a slot, then we need the editor, regardless (in theory) of whether or not there is a bound Content object.

Also removed a deprecated (and not-used-in-core) method on sfContext.