[Editor] Fixing bug with switching from slot type to slot type.
To get this switching to work, the entire sympalSlot widget needs to be replaced. That's because the element that we apply the sympalSlot to (.sympal_slot_wrapper) contains metadata that changes when the type changes. Thinking about it now, it may be possible to do this without refreshing the whole thing but just by changing that metadata...

The change type action responds with the entire slot editor and javascript to reinitialize all of the sympalSlots and to re-open the current one.