[installer] Reworking the install - a work in progress.
The installer should only do things that are normally expected of developers (as a service). Normally, we'd expect a developer to install symfony, download the sympal plugin, and setup the databases.yml file. The install is a tool to automate those tasks and those tasks alone. Any tasks that are specific to installing sympal or making any modifications needed by sympal should be done in the sympal:install task. That way, if a seasoned developer installs symfony and downloads sympal, they can still use the sympal-specific install features.