[task] Adding phpDoc to create/delete site task. Removed a bunch of user options from create-site.
The users, as I see them currently, are global. This may need to change later, but a user on one site is a user on another site.

For that reason, the creation of the user fixtures is more global that creating an app - those fixtures are copied when you create the whole project.