[install] The install script now copies all data/fixtures/project/*.yml.sample files from each sympal plugin into the data/fixture/sympal directory.
The idea is that each plugin ships with NO fixtures yml file that is automatically included. Rather, each ships with a sample file that is copied to the core directory where it can be used.

Furthermore, the fixtures yaml is being organized into two different categories: project-wide fixtures and site-wide fixtures. This first pass is the project fixtures, the site fixtures are up next.