Fixing strange bug with route generation and custom module and action.
The strange thing is that I'm not sure what the offending code was intended to do, which is what was removed. In the calculation of the path of a Content object, you have the following tree of priorities:
  1) Does the sfSympalContent record have a custom_path value?
  2) Does the Content Type object specify a route pattern (e.g. /blogs/:slug)?
  3) Default to /content/:slug

Before this commit, there was special handling between 1 and 2 for custom module and action. I don't understand what this did nor what might be affected by its removal. But, everything appears to be right with the world.