[routing] Follow up to sha:43f06f6 - handles a bug with generating routes for Content records with custom module or action.
The tricky part is that these situations merit the generation of their own route. That's fine, except that we need to be able to determin the full, unpatternized url to that object.

My approach to this will work fine, but may be slow (depending on how much of this is eventually cached). In the generation of the route, I create a second, temporary route, to see what the full url of the route would have been had there been no custom module and action.