[install] The final piece of the new install procedure, which relates to test fixtures.
There are no longer any .yml files that live in the plugins that are loaded. Instead, fixtures are copied:
  1) On sympal:install, any fixtures in a plugin's data/fixtures/project are copied to the project's data/fixtures/sympal directory
  2) On site creation, any fixtures in a plugin's data/fixtures/site are copied to the project's data/fixtures/sympal/SITE_NAME directory

For upgrades, re-running sympal:install should be enough to copy these assets.

These assets are copied each time you run one of the above tasks, but the point is that those tasks should only be run once, not repeatedely as before.

sympal:install is no longer a replacement for doctrine:build - sympal:install should be run once, doctrine:build should be used after that.