[isntall] Effectively disabling any action when installing plugins.
Though the transformation of this process is still being thought out, the idea is that everything happens organically. Test fixtures are copied into the project and between that and the schema, all the installation needed is taken care of.

There are a few tasks that we will need to offload to the plugin install, that are conveniently taken care of here during the main install including:
 * publishing plugin assets
 * copying fixtures
 * eventual migrations