[content list] fixing bug with custom dql queries
If there was custom DQL query specified for content list record
PluginsfSympalContentList::_buildQuery() executes Doctrine_Query::query()
which parsesDqlQuery and then executes it returning Doctrine_Collection.
Which is obviously wrong, because we need only to build Doctrine_Query object
in order to pass it further to the pager.

Adding @return phpDoc to _buildQuery()